+++ src/modules/manufacturing/entities/gis-header.ts
@@ -16,9 +16,9 @@ export class GisHeader extends BaseEntity {
   @Column({ name: 'REPORT_DATE' })
   reportDate!: Date;
 
-  @Field()
+  @Field({ nullable: true })
   @Column({ name: 'ORDER_NO' })
-  orderNo!: string;
+  orderNo?: string;
 
   @Field()
   @Column({ name: 'PART_NO' })
@@ -81,12 +81,20 @@ export class GisHeader extends BaseEntity {
   seriBeam?: string;
 
   @Field({ nullable: true })
-  @Column({ name: 'ROLL' })
-  roll?: number;
+  @Column({ name: 'ROLL_NO' })
+  rollNo?: number;
 
   @Field({ nullable: true })
   @Column({ name: 'INSPECT_TYPE' })
   inspectType?: string;
+
+  @Field({ nullable: true })
+  @Column({ name: 'LOT_BATCH_NO' })
+  lotBatchNo?: string;
+
+  @Field({ nullable: true })
+  @Column({ name: 'NOTES' })
+  notes?: string;
 
   @Field()
   @Column({ name: 'CREATED_BY' })